map data bug fix

+++ raspi/localization/wifi_estimate.cpp
@@ -71,6 +71,7 @@ vector<PointEstimate> WiFiEstimate::ClosestByMahalanobis(const vector<Result>& s
   // Mahalanobis distance.
   for (auto& node : current_likely_points) {
     auto point = node->point;
+    //printf("checking likely point: x = %.1f, y = %.1f", point->x, point->y);
     if (debug and distance(point->x, point->y, realx - Global::FilterBiasX,
         realy - Global::FilterBiasY) > Global::FilterableDistance) {
       continue;
@@ -106,6 +107,25 @@ vector<PointEstimate> WiFiEstimate::ClosestByMahalanobis(const vector<Result>& s
               point, df, sqrt(sum)));
     }
   }
+
+//  DEBUG CODE
+//  char buffer[100];
+//  sprintf(buffer, "\n%7s %7s %7s %7s %7s\n",
+//      "weight", "x", "y", "df", "metric");
+//  cout << buffer;
+//  sort(point_stats.begin(), point_stats.end(),
+//      [](const stat &a, const stat &b) -> bool {
+//          return get<0>(a) > get<0>(b);
+//      });
+//  for (auto&& p_stat : point_stats) {
+//    sprintf(buffer, "%7.2e %7.1f %7.1f %7d %7.2e\n",
+//        get<0>(p_stat),
+//        get<1>(p_stat)->x,
+//        get<1>(p_stat)->y,
+//        get<2>(p_stat),
+//        get<3>(p_stat));
+//    cout << buffer;
+//  }
 
 //  if (v & WIFI_VARIANT_TOP1 or v & WIFI_VARIANT_TOP_FEW) {
 //    sort(point_stats.begin(), point_stats.end(),
@@ -319,7 +339,7 @@ vector<Result> AverageScans(vector<vector<Result>> scans) {
   for (auto& kv : macs) {
     if (kv.second.size() < min_count)
       continue;
-    Result r = { kv.first, (int)mean(kv.second) };
+    Result r = { kv.first, mean(kv.second) };
     averaged.push_back(r);
   }
   return averaged;
